# docs/swagger.yaml
definitions:
...
  routes.MessageResponse:
    properties:
      code:
        type: integer
      details:
        type: string
    type: object
...
  routes.institutionCreateParams:
    properties:
      country:
        type: string
      dec:
        type: integer
      description:
        type: string
      dni:
        type: string
      email:
        type: string
      name:
        type: string
      nemo:
        type: string
      owner:
        type: string
      state:
        type: integer
      verified:
        type: string
    required:
    - country
    - dni
    - email
    - owner
    type: object
  routes.institutionUpdateParams:
    properties:
      country:
        type: string
      dec:
        type: integer
      description:
        type: string
      email:
        type: string
      owner:
        type: string
      state:
        type: integer
    required:
    - country
    type: object
...
  routes.responseAudit:
    properties:
      data: {}
    type: object
  routes.responseInstitution:
    properties:
      data:
        description: Data data.Sensor `json:"data"`
    type: object
...
info:
  contact: {}
  description: api autentia manager
  title: Autentia Manager Documentation
  version: "1"
paths:
  /autentia/audits/{code}:
    get:
      consumes:
      - application/json
      description: get autentia audits
      parameters:
      - description: code
        in: path
        name: code
        required: true
        type: string
      - description: country
        in: query
        name: country
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/routes.responseAudit'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/routes.MessageResponse'
      security:
      - BarerToken: []
      summary: get autentia audits
      tags:
      - autentia audits
  /autentia/institutions:
    get:
      consumes:
      - application/json
      description: get autentia institutions
      parameters:
      - description: country
        in: query
        name: country
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/routes.responseInstitution'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/routes.MessageResponse'
      security:
      - BarerToken: []
      summary: get autentia institutions
      tags:
      - autentia institutions
    post:
      consumes:
      - application/json
      description: create institution
      parameters:
      - description: institution
        in: body
        name: institution
        required: true
        schema:
          $ref: '#/definitions/routes.institutionCreateParams'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/routes.responseInstitution'
        "406":
          description: Not Acceptable
          schema:
            $ref: '#/definitions/routes.MessageResponse'
      security:
      - BarerToken: []
      summary: create institution
      tags:
      - autentia institutions
  /autentia/institutions/{name}:
    delete:
      consumes:
      - application/json
      description: update institution
      parameters:
      - description: name
        in: path
        name: name
        required: true
        type: string
      - description: country
        in: query
        name: country
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/routes.responseInstitution'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/routes.MessageResponse'
      security:
      - BarerToken: []
      summary: update institution
      tags:
      - autentia institutions
    get:
      consumes:
      - application/json
      description: get autentia institution
      parameters:
      - description: name
        in: path
        name: name
        required: true
        type: string
      - description: country
        in: query
        name: country
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/routes.responseInstitution'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/routes.MessageResponse'
      security:
      - BarerToken: []
      summary: get autentia institution
      tags:
      - autentia institutions
    put:
      consumes:
      - application/json
      description: update institution
      parameters:
      - description: name
        in: path
        name: name
        required: true
        type: string
      - description: institution
        in: body
        name: institution
        required: true
        schema:
          $ref: '#/definitions/routes.institutionUpdateParams'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/routes.responseInstitution'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/routes.MessageResponse'
      security:
      - BarerToken: []
      summary: update institution
      tags:
      - autentia institutions
  /autentia/institutions/all:
    get:
      consumes:
      - application/json
      description: get all autentia institutions
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/routes.responseInstitution'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/routes.MessageResponse'
      security:
      - BarerToken: []
      summary: get all autentia institutions
      tags:
      - autentia institutions
...
securityDefinitions:
  BarerToken:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"